backend/requirements.txt
# Flask & Extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# PDF Manipulation
PyPDF2==3.0.1
reportlab==4.0.7

# Image Processing
Pillow==10.1.0

# Background Removal (Optional - comment out if installation fails)
# rembg requires additional dependencies and ~1GB download on first run
# If not installed, falls back to simple threshold-based removal
rembg==2.0.50

# Encryption
cryptography==41.0.7

# Database
Flask-SQLAlchemy==3.1.1

# Environment
python-dotenv==1.0.0
